+++ python/GafferUI/Menu.py
@@ -371,7 +371,9 @@ class Menu( GafferUI.Widget ) :
 		if item.checkBox is not None :
 			qtAction.setCheckable( True )
 			checked = self.__evaluateItemValue( item.checkBox )
-			qtAction.setChecked( checked )
+			# if a callable was passed as the "checkBox" parameter it may return None
+			if checked is not None :
+				qtAction.setChecked( checked )
 
 		if item.command :